Move FlowChart tests into test folder
