+++ packages/builder/src/components/automation/AutomationBuilder/FlowChart/test/AutomationStepHelpers.test.ts
@@ -7,7 +7,7 @@ import {
   branchStep,
   nestedLoopBranchAutomation,
 } from "@/test/automationFixtures"
-import { getLogStepData, processLogSteps } from "./AutomationStepHelpers"
+import { getLogStepData, processLogSteps } from "../AutomationStepHelpers"
 
 vi.mock("@/stores/builder", () => {
   return {
